requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

langchain==0.3.12
langchain-community==0.3.12
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
chromadb==0.5.23

pypdf==5.1.0

# MCP (Model Context Protocol) - Python SDK
mcp==1.2.0
anyio==4.7.0
